% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tapplyAggregateMethod.R
\name{ds.tapplyAggregateMethod}
\alias{ds.tapplyAggregateMethod}
\title{Apply a DataSHIELD aggregate function by groups across servers}
\usage{
ds.tapplyAggregateMethod(
  X.name,
  INDEX.names = NULL,
  FUN.name,
  FUN.pars = list(),
  checks = FALSE,
  simplify = FALSE,
  array2DF = FALSE,
  wrapErrors = TRUE,
  datasources = datashield.connections_find()
)
}
\arguments{
\item{X.name}{Character string specifying the name of the server-side object
to be summarized, for example \code{"pooled$weight_T0"}.}

\item{INDEX.names}{Character vector specifying one or more server-side
grouping variables, for example \code{"pooled$sex"} or
\code{c("pooled$sex", "pooled$country")}. If \code{NULL}, no grouping
variable is used.}

\item{FUN.name}{Character string specifying the name of the server-side
aggregate function to apply, for example \code{"boxPlotGGDS"}.}

\item{FUN.pars}{A list of additional parameters passed to the server-side
aggregate function. Default is an empty list.}

\item{checks}{Logical. If \code{TRUE}, check that \code{X.name} and
\code{INDEX.names} are defined on all servers and refer to atomic objects
or factors. Default is \code{FALSE}.}

\item{simplify}{Logical passed to \code{tapplyAggregateMethodDS}. If
\code{TRUE}, simplify grouped output where possible. Default is
\code{FALSE}.}

\item{array2DF}{Logical passed to \code{tapplyAggregateMethodDS}. If
\code{TRUE}, convert array-like output to a data frame where supported.
Default is \code{FALSE}.}

\item{wrapErrors}{Logical passed to \code{tapplyAggregateMethodDS}. If
\code{TRUE}, server-side errors are wrapped and returned rather than
stopping immediately. Default is \code{TRUE}.}

\item{datasources}{A list of DataSHIELD connections. Each element must be of
class \code{DSConnection}. Default is
\code{datashield.connections_find()}.}
}
\value{
A list containing one result per DataSHIELD server, as returned by
  \code{DSI::datashield.aggregate}.
}
\description{
Calls the server-side function \code{tapplyAggregateMethodDS} on each
DataSHIELD server. The function applies an aggregate server-side function
\code{FUN.name} to \code{X.name}, optionally stratified by one or more
grouping variables given in \code{INDEX.names}.
}
\details{
This is a client-side wrapper around \code{DSI::datashield.aggregate}. It is
useful when the same aggregate function should be evaluated by groups on all
connected DataSHIELD servers.
}
\examples{
if (require(DSLite) && require(dsBase)) {
data('CNSIM1')
data('CNSIM2')

# build a DSLite server with the datasets inside
dslite.server1 <- DSLite::newDSLiteServer(tables=list(table1=CNSIM1))
dslite.server2 <- DSLite::newDSLiteServer(tables=list(table2=CNSIM2))

#dslite.server1$aggregateMethod("boxPlotGGDS", dsBase::boxPlotGGDS)
#dslite.server2$aggregateMethod("boxPlotGGDS", dsBase::boxPlotGGDS)

# build DS login information
builder <- DSI::newDSLoginBuilder()
builder$append(server = 'server1', driver = 'DSLiteDriver', url = 'dslite.server1')
builder$append(server = 'server2', driver = 'DSLiteDriver', url = 'dslite.server2')
logindata <- builder$build()

# do login and table assignment
cur_conn <- DSI::datashield.login(logindata)
datashield.assign.table(cur_conn, 'pooled', table = list(server1='table1', server2='table2'))

# unstratified
ds.tapplyAggregateMethod(
  X.name = "pooled$LAB_TRIG",
  FUN.name = "meanDS",
  checks = FALSE,
  array2DF = T,
  datasources = cur_conn
)

# stratified
ds.tapplyAggregateMethod(
  X.name = "pooled$LAB_TRIG",
  INDEX.names = "pooled$GENDER",
  FUN.name = c("meanDS","varDS"),
  checks = FALSE,
  array2DF = T,
  datasources = cur_conn
)

res <- ds.tapplyAggregateMethod(
  X.name = "pooled$LAB_TRIG",
  INDEX.names = "pooled$GENDER",
  FUN.name = "meanDS",
  checks = FALSE,
  datasources = cur_conn
)
#poolMeanDS(res) gives wrong results but should return an error, because the structure is not as expected
poolDS(res)

lapply( res, function(x) {lapply(x, array2DF)}) # same output as ds.tapplyAggregateMethod with array2DF = TRUE

res <- ds.tapplyAggregateMethod(
  X.name = "pooled$LAB_TRIG",
  INDEX.names = "pooled$GENDER",
  FUN.name = "boxPlotGGDS",
  checks = FALSE,
  datasources = cur_conn
)  # ERROR?
}

}
