% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolHistogramDS1.R
\name{histlist2plot}
\alias{histlist2plot}
\title{Plot a list of DataSHIELD histogram results}
\usage{
histlist2plot(hist_list, ycounts = FALSE, xlabel = "mids", title = "title")
}
\arguments{
\item{hist_list}{A non-empty list. Each element should contain a
\code{histobject} component with standard histogram elements:
\code{breaks}, \code{counts}, \code{density}, and \code{mids}.}

\item{ycounts}{Logical. If \code{TRUE}, plot bin counts on the y-axis.
If \code{FALSE}, plot densities. Default is \code{FALSE}.}

\item{xlabel}{Character string used as the x-axis label. Default is
\code{"mids"}.}

\item{title}{Character string used as the legend title and, currently, the
x-axis label. Default is \code{"title"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Converts a list of server-side histogram outputs into a stacked
\pkg{ggplot2} bar plot. The plot can display either bin counts or densities.
}
\examples{
data1 <- rnorm(100)
data2 <- rnorm(100)
xlim <- range(c(data1,data2))
xlim[1] <- floor(xlim[1])
xlim[2] <- ceiling(xlim[2])
breaks <- seq( xlim[1], xlim[2], length.out=10)

h1 <- hist(data1, breaks = breaks, plot = FALSE)
h2 <- hist(data2, breaks = breaks, plot = FALSE)

invalidcells1 <- (h1$counts < 3) & (h1$counts > 0)
invalidcells2 <- (h2$counts < 3) & (h2$counts > 0)
h1$counts[invalidcells1] <- 0#NA
h2$counts[invalidcells2] <- 0#NA
h1$density[invalidcells1] <- 0#NA
h2$density[invalidcells2] <- 0#NA

res <- list(
  server1 = list(histobject = h1, invalidcells = sum(invalidcells1)),
  server2 = list(histobject = h2, invalidcells = sum(invalidcells2))
)
class(res) <- c("dsPooledHistogram", class(res))

histlist2plot(res)

}
