% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolHistogramDS1.R
\name{plot.dsPooledHistogram}
\alias{plot.dsPooledHistogram}
\title{Plot pooled DataSHIELD histogram results}
\usage{
\method{plot}{dsPooledHistogram}(
  x,
  freq = FALSE,
  xlab = "Value",
  main = "Histogram of the pooled data",
  ...
)
}
\arguments{
\item{x}{An object of class \code{dsPooledHistogram}, as returned by
\code{poolHistogramDS2}.}

\item{freq}{Logical. If \code{TRUE}, plot counts. If \code{FALSE}, plot
densities. Default is \code{FALSE}.}

\item{xlab}{Character string for the x-axis label. Default is
\code{"Value"}.}

\item{main}{Character string for the plot title. Default is
\code{"Histogram of the pooled data"}.}

\item{...}{Further arguments passed to \code{graphics::plot.histogram}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Creates a histogram plot from an object returned by \code{poolHistogramDS2}.
}
\examples{
data1 <- rnorm(100)
data2 <- rnorm(100)
xlim <- range(c(data1, data2))
breaks <- seq(floor(xlim[1]), ceiling(xlim[2]), length.out = 10)

h1 <- hist(data1, breaks = breaks, plot = FALSE)
h2 <- hist(data2, breaks = breaks, plot = FALSE)

res <- list(
  server1 = list(histobject = h1, invalidcells = 0L),
  server2 = list(histobject = h2, invalidcells = 0L)
)

pooled <- poolHistogramDS2(res)
plot(pooled)

}
