% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolHistogramDS1.R
\name{poolHistogramDS2}
\alias{poolHistogramDS2}
\title{Pool DataSHIELD histogram results across servers}
\usage{
poolHistogramDS2(outputs)
}
\arguments{
\item{outputs}{A non-empty list. Each element should be the server-side
histogram result and must contain \code{histobject} and \code{invalidcells}.
The \code{histobject} should contain standard histogram components such as
\code{breaks}, \code{counts}, \code{density}, and \code{mids}. All histogram
objects are assumed to use identical \code{breaks} and \code{mids}.}
}
\value{
A named list with elements:
\describe{
  \item{histObject}{Combined histogram object with pooled \code{counts},
    \code{density}, and \code{intensities}.}
  \item{invalidCells}{Summary of invalid cell counts across servers.}
}
}
\description{
Pools server-side histogram outputs by summing bin counts across studies and
combining densities for histograms constructed using common break points.
Each server result is expected to contain a histogram object and invalid cell
information.
}
\examples{
data1 <- rnorm(100)
data2 <- rnorm(100)
xlim <- range(c(data1,data2))
xlim[1] <- floor(xlim[1])
xlim[2] <- ceiling(xlim[2])
breaks <- seq( xlim[1], xlim[2], length.out=10)

h1 <- hist(data1, breaks = breaks, plot = FALSE)
h2 <- hist(data2, breaks = breaks, plot = FALSE)

invalidcells1 <- (h1$counts < 3) & (h1$counts > 0)
invalidcells2 <- (h2$counts < 3) & (h2$counts > 0)
h1$counts[invalidcells1] <- 0#NA
h2$counts[invalidcells2] <- 0#NA
h1$density[invalidcells1] <- 0#NA
h2$density[invalidcells2] <- 0#NA

res <- list(
  server1 = list(histobject = h1, invalidcells = sum(invalidcells1)),
  server2 = list(histobject = h2, invalidcells = sum(invalidcells2))
)

pooled <- poolHistogramDS2(res)
graphics::plot(pooled$histObject, freq = FALSE,
  xlab="my  data", main='Histogram of the pooled data')

}
