% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolBoxPlotDS.R
\name{preprocessBoxPlot}
\alias{preprocessBoxPlot}
\title{Preprocess numeric data for DataSHIELD box plots}
\usage{
preprocessBoxPlot(vector, datasources = datashield.connections_find())
}
\arguments{
\item{vector}{Character string specifying the server-side numeric vector to
preprocess, for example \code{"pooled$weight_T0"}.}

\item{datasources}{DataSHIELD connections object. Default is
\code{datashield.connections_find()}.}
}
\value{
Invisibly returns the result of \code{DSI::datashield.assign.expr}.
  The main effect is assignment of \code{boxPlotRawDataNumeric} on each
  server.
}
\description{
Calls the server-side preprocessing function
\code{boxPlotGG_data_Treatment_numericDS} for a numeric vector and assigns
the result on each DataSHIELD server as \code{boxPlotRawDataNumeric}.
}
\examples{
if (require(DSLite)) {
data('CNSIM1')
data('CNSIM2')

# build a DSLite server with the datasets inside
dslite.server1 <- DSLite::newDSLiteServer(tables=list(table1=CNSIM1))
dslite.server2 <- DSLite::newDSLiteServer(tables=list(table2=CNSIM2))

# build DS login information
builder <- DSI::newDSLoginBuilder()
builder$append(server = 'server1', driver = 'DSLiteDriver', url = 'dslite.server1')
builder$append(server = 'server2', driver = 'DSLiteDriver', url = 'dslite.server2')
logindata <- builder$build()

# do login and table assignment
cur_conn <- DSI::datashield.login(logindata)
datashield.assign.table(cur_conn, 'pooled', table = list(server1='table1', server2='table2'))

preprocessBoxPlot("pooled$PM_BMI_CONTINUOUS", datasources = cur_conn)
cally <- "boxPlotGGDS(boxPlotRawDataNumeric, NULL, NULL)"
DSI::datashield.aggregate(cur_conn, as.symbol(cally))

preprocessBoxPlot("pooled$LAB_TSC", datasources = cur_conn)
cally <- "boxPlotGGDS(boxPlotRawDataNumeric, pooled$GENDER, pooled$PM_BMI_CATEGORICAL)"
srv_res <- DSI::datashield.aggregate(cur_conn, as.symbol(cally))
}

}
