% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolTable1DDS.R
\name{print.dsPooledTable1D}
\alias{print.dsPooledTable1D}
\title{Print a pooled DataSHIELD 1D table}
\usage{
\method{print}{dsPooledTable1D}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dsPooledTable1D}, as returned by
\code{poolTable1DDS}.}

\item{...}{Further arguments passed to \code{print}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints the combined (pooled) counts and percentages from an object of class
\code{dsPooledTable1D}. This method is intended to provide a concise summary
of the pooled result.
}
\examples{
\dontrun{
res <- poolTable1DDS(pool_res)
print(res)
}

}
